fixes dusting 2 (#22749)

+++ code/modules/mob/living/carbon/carbon.dm
@@ -10,6 +10,8 @@
 		qdel(I)
 	// This clause is here due to items falling off from limb deletion
 	for(var/obj/item in get_all_slots())
+		if(QDELETED(item))
+			continue
 		unEquip(item, silent = TRUE)
 		qdel(item)
 	QDEL_LIST_CONTENTS(internal_organs)
